Add pubkey auth to nss endpoints
